+++ src/main/java/jacchm/footballapp/controller/TeamsController.java
@@ -3,7 +3,6 @@ package jacchm.footballapp.controller;
 import jacchm.footballapp.mapping.dto.TeamDTO;
 import jacchm.footballapp.service.TeamsService;
 import lombok.RequiredArgsConstructor;
-import lombok.extern.slf4j.Slf4j;
 import org.springframework.web.bind.annotation.DeleteMapping;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PathVariable;
@@ -12,7 +11,6 @@ import org.springframework.web.bind.annotation.RestController;
 
 import java.util.List;
 
-@Slf4j
 @RequiredArgsConstructor
 @RestController
 @RequestMapping("/teams")
@@ -20,7 +18,7 @@ public class TeamsController {
 
     private final TeamsService teamsService;
 
-    // TODO: method below should be only used by Admins
+    // TODO: scheduled or done manually by admin
     @GetMapping("/updateAll")
     public boolean updateAll() {
        return teamsService.updateAll();
@@ -37,7 +35,10 @@ public class TeamsController {
         return teamsService.getAllLeagueTeams(id);
     }
 
-
+    @GetMapping("/getTeam/{id}")
+    public TeamDTO getTeam(@PathVariable("id") Integer id){
+        return teamsService.getTeam(id);
+    }
 
 }
